完成Tags搜索功能，优化设置界面,逻辑优化

+++ yande.re/tagsDelegate.m
@@ -2,8 +2,8 @@
 //  TagsDelegate.m
 //  yande.re
 //
-//  Created by hnzc on 16/5/18.
-//  Copyright © 2016年 hnzc. All rights reserved.
+//  Created by YuJie on 16/5/18.
+//  Copyright © 2016年 YuJie. All rights reserved.
 //
 
 #import "TagsDelegate.h"
@@ -44,7 +44,7 @@
     for (NSDictionary *temp in tagSource) {
         NSString *tagName=[temp valueForKey:@"name"];
         NSUInteger width=tagName.length*11;
-        [widthArray addObject:[NSString stringWithFormat:@"%lu",width]];
+        [widthArray addObject:[NSString stringWithFormat:@"%lu",(unsigned long)width]];
     }
 }
